+++ src/lib/table/Cell.ts
@@ -1,13 +1,9 @@
 export default class Cell {
-  private _content: string;
+  content: string;
   width: number;
 
   constructor(content: string = "") {
-    this._content = content.trim();
-    this.width = this._content.length + 2;
-  }
-
-  formattedContent(width: number) {
-    return ` ${this._content} `.padEnd(width);
+    this.content = content.trim();
+    this.width = this.content.length + 2;
   }
 }