+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Command and Context menu to include/exclude an header from TOC
 - TOC commands are now accessible through context menu
 
 ## [0.0.3] - 2024-02-04